move (now generalized) expression matching to util
